spot_sdk_interface/nlmap_spot/requirements.txt
absl-py==1.4.0
addict==2.4.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
cachetools==4.2.4
# certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
click==8.0.4
clip @ git+https://github.com/openai/CLIP.git@3702849800aa56e2223035bccd1c6ef91c704ca8
comm==0.1.2
ConfigArgParse==1.5.3		
#contourpy==1.0.7
cycler==0.11.0
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.4
easydict==1.10
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.4.1
Flask==2.0.3
flatbuffers==23.1.21
# flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.27.1
ftfy==6.0.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==6.0.0a0
ipython==7.16.3
ipywidgets==8.0.0b0
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==3.0.3

scikit-learn==0.24.2
scipy==1.5.4
Pillow==8.4.0
open3d==0.15.2
opencv-python==4.7.0.68
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.21.0
torch==1.7.1
torchvision==0.8.2
virtualenv==20.13.2


joblib
jsonschema
jupyter_client
jupyter_core
jupyterlab-widgets
keras
kiwisolver
libclang
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mkl-fft
#mkl-service
nbformat
nest-asyncio
numpy
oauthlib
opt-einsum
packaging
pandas
parso
pexpect
pickleshare
pkgutil_resolve_name
platformdirs
plotly
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
pyasn1
pyasn1-modules
Pygments
pyparsing
pyquaternion
pyrsistent
python-dateutil
pytz
PyYAML
pyzmq
regex
requests
requests-oauthlib
rsa
stack-data
tenacity
termcolor
threadpoolctl
tornado
tqdm
traitlets
urllib
wcwidth
Werkzeug
widgetsnbextension
wrapt
zipp
